v1/backend/pyproject.toml
[project]
name = "interview-ai-backend"
version = "0.1.0"
description = "Real-time AI interview preparation chatbot backend"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[daily,openai,deepgram,silero]>=0.0.98",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.0",
    "websockets>=13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
